Remove books_action_types.js books_action_creators.js and books_state.js files
