--- src/redux/books/books_action_types.js
@@ -1,2 +1,0 @@
-export const CREATE_BOOK = 'bookStore/books/CREATEBOOK';
-export const REMOVE_BOOK = 'bookStore/books/REMOVEBOOK';
