--- src/redux/books/books_state.js
@@ -1,5 +1,0 @@
-import booksArray from '../../data/index';
-
-const initialState = booksArray;
-
-export default initialState;
